--output for TC + deserialization for recursion
