+++ mod-tool/src/main/scala/at/forsyte/apalache/tla/tooling/opt/TypeCheckCmd.scala
@@ -15,5 +15,7 @@ class TypeCheckCmd
   var file: File = arg[File](description = "a TLA+ specification (.tla or .json)")
   var inferPoly: Boolean = opt[Boolean](name = "infer-poly", default = true,
       description = "allow the type checker to infer polymorphic types, default: true")
+  var output: Option[String] = opt[Option[String]](name = "output",
+      description = "file to which the typechecked source is written (.tla or .json), default: None")
 
 }
