+++ tla-io/src/main/scala/at/forsyte/apalache/tla/imp/passes/SanyParserPassImpl.scala
@@ -9,7 +9,7 @@ import at.forsyte.apalache.tla.lir.{CyclicDependencyError, TlaModule}
 import at.forsyte.apalache.tla.lir.storage.{ChangeListener, SourceLocator}
 import at.forsyte.apalache.tla.lir.transformations.standard.DeclarationSorter
 import at.forsyte.apalache.io.lir.{TlaWriter, TlaWriterFactory}
-import at.forsyte.apalache.tla.imp.{SanyImporter, SanyImporterException}
+import at.forsyte.apalache.tla.imp.{SanyImporter, SanyImporterException, utils}
 import com.google.inject.Inject
 import com.google.inject.name.Named
 import com.typesafe.scalalogging.LazyLogging
@@ -73,41 +73,7 @@ class SanyParserPassImpl @Inject() (
         writerFactory.writeModuleAllFormats(rootModule.get.copy(name = "00_OutParser"), TlaWriter.STANDARD_MODULES)
 
         // write parser output to specified destination, if requested
-        options.get[String]("io", "output").foreach { output =>
-          val outfile = new File(output)
-          val outfileName = outfile.toString()
-
-          val ext = FilenameUtils.getExtension(outfileName)
-          val name = FilenameUtils.getBaseName(outfileName)
-
-          ext match {
-            case "tla" =>
-              writerFactory.writeModuleToTla(
-                  rootModule.get.copy(name),
-                  TlaWriter.STANDARD_MODULES,
-                  Some(outfile),
-              )
-            case "json" =>
-              writerFactory.writeModuleToJson(
-                  rootModule.get.copy(name),
-                  TlaWriter.STANDARD_MODULES,
-                  Some(outfile),
-              )
-            case _ =>
-              logger.error(s"  > Unrecognized file format: ${outfile.toString}. Supported formats: .tla and .json")
-          }
-
-          if (options.getOrElse[Boolean]("general", "debug", false)) {
-            val sourceLocator =
-              SourceLocator(sourceStore.makeSourceMap, new ChangeListener())
-            rootModule.get.operDeclarations foreach sourceLocator.checkConsistency
-          }
-          if (options.getOrElse[Boolean]("general", "debug", false)) {
-            val sourceLocator =
-              SourceLocator(sourceStore.makeSourceMap, new ChangeListener())
-            rootModule.get.operDeclarations foreach sourceLocator.checkConsistency
-          }
-        }
+        utils.writeToOutput(rootModule.get, options, writerFactory, logger, sourceStore)
 
         rootModule.map { m => nextPass.updateModule(this, m) }
         true
